// src/layrQuery/types/crudEnums.ts
export enum crudEnums {

    create = "create",
    get = "get",
    update = "update",
    delete = "delete",
    otherQuery = "otherQuery"//ha a savepartsnal csak ugy van egy query,amit futtatni akarsz, akkor azt itt kell rogzíteni. a path segitsegevel ra lehet mutatni, hogy hol jojjon fel lehetosegkent a pathban
}